[FEATURE] Encoder implemented
